port: fix wall-collision teleport — Dot2D/Det2D read 4-byte gte result, not long
HZD_HazardReaction -> Len2D -> Dot2D returned *(long*)(SCRPAD+0x008), but
gte_NormalClip writes only a 4-byte int there. On the 64-bit port `long` is
8 bytes, so Dot2D returned {cross, stale_high_bytes}; the stale high word (the
adjacent {x,z}, often negative) made the 64-bit value negative, so
SquareRoot0() returned 0. That made HZD_HazardReaction see area==0 and bail
without reacting, so CheckNear's retry amplified a degenerate step via
GV_LenVec3 (len=1) into a ~24000-unit vector — Snake's short mov.vz overflowed
and teleported across the map when pushing into a wall.

Read the result as *(int*) (4 bytes) in Dot2D/Det2D. Match-safe: long==int==4
bytes on PSX, identical compiled output. Verified with the
10_collision_merge_regression replay: Snake now slides along the wall.

+++ source/libhzd/private.h
@@ -47,13 +47,13 @@ static inline long Dot2D(SVECTOR *v1, SVECTOR *v2)
     *(short *)(SCRPAD_ADDR + 0x004) = -v2->vy;
     *(short *)(SCRPAD_ADDR + 0x006) = v2->vx;
     gte_NormalClip(0, *(long *)v1, *(long *)(SCRPAD_ADDR + 0x004), (SCRPAD_ADDR + 0x008));
-    return *(long *)(SCRPAD_ADDR + 0x008);
+    return *(int *)(SCRPAD_ADDR + 0x008); /* gte_NormalClip writes a 4-byte int; long is 8 bytes on the 64-bit port and would read stale high bytes */
 }
 
 static inline long Det2D(SVECTOR *v1, SVECTOR *v2)
 {
     gte_NormalClip(0, *(long *)v1, *(long *)v2, (SCRPAD_ADDR + 0x008));
-    return *(long *)(SCRPAD_ADDR + 0x008);
+    return *(int *)(SCRPAD_ADDR + 0x008); /* gte_NormalClip writes a 4-byte int; long is 8 bytes on the 64-bit port and would read stale high bytes */
 }
 
 static inline long Len2D(SVECTOR *vec)
